requirements.txt
aws-cdk-lib==2.70.0
constructs>=10.0.0,<11.0.0
aws-solutions-constructs.aws-lambda-s3>=2.35.0,<3.0.0
